+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+* HEAD
+
+  - Write an empty key to S3 even if nothing is written to S3OpenWrite (PR #61, @petedmarsh)
+
 * 1.3.2, 3rd January 2016
 
   - Bug fix release to enable 'wb+' file mode (PR #50)